davinci: Remove audio-kernel from source
audio-kernel is a part of Techpack module which is updated separately
from kernel. Right now audio-kernel is unused.  To make it actually
usable we need to introduce Techpack support and squash the contents
of audio-kernel there.

For now just remove unused audio code from tree. Later commit will
bring it up again in a proper way.

Signed-off-by: Alesaiko <solcmdr@gmail.com>