techpack: audio: Initial bring up for davinci
Techpack is a kernel module which is updated in a different way than
other kernel parts. Currently its only content is audio related code
like codec support.

To provide support for audio subsystem in davinci device, add techpack
support in the tree and squash audio-kernel in techpack/audio target.

Note that this variant of techpack is different from the one from
Code Aurora Forum. I've stripped the code which is unused on davinci
platform.

Signed-off-by: Alesaiko <solcmdr@gmail.com>